requirements.txt
aiosmtplib==4.0.1
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.13
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
azure-functions
azure-functions-worker
httpx
python-multipart
pydantic[email]